+++ scripts/ts/danger/__tests__/changelog.test.ts
@@ -11,7 +11,7 @@ import {
   singleEpicBpdStory
 } from "../__mocks__/storyMock";
 import { getChangelogScope, getStoryChangelogScope } from "../changelog";
-import { fromPivotalToGenericTicket } from "../story/types";
+import { fromPivotalToGenericTicket } from "../../common/ticket/types";
 
 describe("Test pivotal Utility", () => {
   it("getStoryChangelogScope on a story without labels should return Right none", () => {
